+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 - Application 支持设置父级上下文
 - Application 方法 customize、applicationStartedHook 支持接收命令行参数
+- Pac4j Configuration 支持 Kerberos
+- Pac4j Configuration 支持 OIDC
+- Pac4j Configuration 支持 SAML
 
 
 ### 🔔 变化